requirements.txt
google-cloud-speech==2.13.1
google-cloud-translate==3.7.2
pyaudio